+++ zapping/libvbi/exp-html.c
@@ -18,7 +18,7 @@
  *  Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
  */
 
-/* $Id: exp-html.c,v 1.33 2005-01-31 07:21:34 mschimek Exp $ */
+/* $Id: exp-html.c,v 1.33.2.1 2005-05-31 02:40:33 mschimek Exp $ */
 
 #ifdef HAVE_CONFIG_H
 #  include "config.h"
@@ -192,7 +192,7 @@ extend				(html_instance *	html,
 				 unsigned int		size)
 {
 	char *buffer;
-	unsigned int n;
+	unsigned long n;
 
 	n = (v->end - v->buffer + incr) * size;
