+++ zapping/libvbi/misc.c
@@ -17,7 +17,7 @@
  *  Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
  */
 
-/* $Id: misc.c,v 1.7 2005-01-31 07:09:49 mschimek Exp $ */
+/* $Id: misc.c,v 1.7.2.1 2005-05-31 02:40:33 mschimek Exp $ */
 
 #include <stdio.h>
 #include <stdlib.h>
@@ -101,7 +101,7 @@ _vbi3_asprintf			(char **		dstp,
 				 ...)
 {
 	char *buf;
-	unsigned int size;
+	unsigned long size;
 	int temp;
 
 	assert (NULL != dstp);
@@ -115,7 +115,7 @@ _vbi3_asprintf			(char **		dstp,
 	for (;;) {
 		va_list ap;
 		char *buf2;
-		int len;
+		long len;
 
 		if (!(buf2 = realloc (buf, size)))
 			break;
@@ -129,7 +129,7 @@ _vbi3_asprintf			(char **		dstp,
 		if (len < 0) {
 			/* Not enough. */
 			size *= 2;
-		} else if ((unsigned int) len < size) {
+		} else if ((unsigned long) len < size) {
 			*dstp = buf;
 			errno = temp;
 			return len;
