+++ zapping/plugins/deinterlace/DI_Misc/DI_OldGame.c
@@ -1,5 +1,5 @@
 /////////////////////////////////////////////////////////////////////////////
-// $Id: DI_OldGame.c,v 1.3.2.2 2005-05-20 05:45:14 mschimek Exp $
+// $Id: DI_OldGame.c,v 1.3.2.3 2005-05-31 02:40:34 mschimek Exp $
 /////////////////////////////////////////////////////////////////////////////
 // Copyright (c) 2001 Lindsey Dubb.  All rights reserved.
 // based on OddOnly and Temporal Noise DScaler Plugins
@@ -20,6 +20,9 @@
 // CVS Log
 //
 // $Log: not supported by cvs2svn $
+// Revision 1.3.2.2  2005/05/20 05:45:14  mschimek
+// *** empty log message ***
+//
 // Revision 1.3.2.1  2005/05/05 09:46:01  mschimek
 // *** empty log message ***
 //
@@ -145,9 +148,9 @@ SIMD_NAME (OldGameFilter)	(TDeinterlaceInfo *	pInfo)
 	if (((unsigned int) pInfo->Overlay |
 	     (unsigned int) pInfo->PictureHistory[0]->pData |
 	     (unsigned int) pInfo->PictureHistory[1]->pData |
-	     (unsigned int) pInfo->OverlayPitch |
-	     (unsigned int) pInfo->InputPitch |
-	     (unsigned int) pInfo->LineLength) & 15)
+	     (unsigned long) pInfo->OverlayPitch |
+	     (unsigned long) pInfo->InputPitch |
+	     (unsigned long) pInfo->LineLength) & 15)
 	    return DeinterlaceOldGameFilter_SSE (pInfo);
     }
